pyproject.toml
[project]
name = "medboxz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.122.0",
    "jinja2>=3.1.6",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "pymongo>=4.15.4",
    "python-jose>=3.5.0",
    "redis>=7.1.0",
    "streamlit>=1.51.0",
    "tavily>=1.1.0",
    "uvicorn>=0.38.0",
]
